身長

+++ db/migrate/20210704045619_create_infos.rb
@@ -1,10 +1,10 @@
 class CreateInfos < ActiveRecord::Migration[6.0]
   def change
     create_table :infos do |t|
-      t.integer :weight,       null:false 
-      t.integer :height,       null:false  
-      t.integer :bmi               
-      t.integer :ave_wei     
+      t.float :weight,       null:false 
+      t.float :height,       null:false  
+      t.float  :bmi               
+      t.float  :ave_wei     
       t.references :user,     foreign_key: true
       t.timestamps
     end
